+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.2 2020-04-15]
+### Added
+- Cut cards to the shoe implementation
+- Burn cards to baccarat game engine.  Will draw top card and burn the appropriate number of cards.
+### Changed
+- Changed RNG shuffle algorithm.  Ensure baccarat decks are initially shuffled.
+
 ## [1.1.1] - 2020-03-12
 ### Added
 - Added BaccaratGameEngine to main module exports